Update Android 4.3+ version of vold-set-sdcard for 4.4 changes
In 4.4, vold-managed volumes no longer set a mountpoint in the fstab --
vold derives the mountpoints itself from the volume's tag.  Change the
script to find the desired entry by vold tag instead (this approach also
works on 4.3).

Users should note that the arguments to vold-set-sdcard have changed --
you need to pass in the vold tag as well as the block device name.